pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "praat-parselmouth>=0.4.7",
    "textgrid>=1.6.1",
]
